implemented StickerType in GQL, moved type defs into own file, fixed authentication
