+++ graphql.js
@@ -11,59 +11,10 @@ const User = require('./src/resolvers/User');
 const Location = require('./src/resolvers/Location');
 const Sticker = require('./src/resolvers/Sticker');
 
+const typeDefs = require('./src/type-defs').typeDefs;
+
 const prisma = new PrismaClient();
 const pubsub = new PubSub();
-
-const typeDefs = gql`
-    scalar Upload
-
-    type Query {
-        stickers(filter: String, skip: Int, take: Int): [Sticker]!
-        collectedStickers: [Sticker]!
-    }
-
-    type Mutation {
-        signup(email: String!, password: String!, name: String!): AuthPayload
-        login(email: String!, password: String!): AuthPayload
-
-        refreshToken(token: String!): AuthPayload
-
-        post(name: String!, lat: Float!, lng: Float!, image: Upload): Sticker!
-        scan(stickerId: ID!): User!
-        deleteSticker(id: ID!): Sticker!
-    }
-
-    type Sticker {
-        id: ID!
-        name: String!
-        location: Location!
-        createdBy: User!
-        createdAt: String
-        imageUrl: String
-    }
-
-    type Location {
-        id: ID!
-        lat: Float!
-        lng: Float!
-    }
-
-    type User {
-        id: ID!
-        name: String!
-        email: String!
-        collectedStickers: [Sticker]!
-    }
-
-    type AuthPayload {
-        token: String
-        user: User
-    }
-
-    type Subscription {
-        newSticker: Sticker
-    }
-`;
 
 const resolvers = {
     Query,
@@ -79,12 +30,13 @@ const server = new ApolloServer({
     resolvers,
     introspection: true,
     playground: true,
-    context: ({ req }) => {
+    context: ({ express: { req } }) => {
         return {
             ...req,
             prisma,
             pubsub,
-            userId: req && req.headers.authorization ? getUserId(req) : null,
+            userId:
+                req && req.headers.authorization ? getUserId(req) : undefined,
         };
     },
 });
